# ui/src/openapi/mifune.openapi.yaml
{
  "openapi" : "3.0.3",
  "info" : {
    "title" : "Mifune API",
    "version" : "0.1.0-SNAPSHOT"
  },
  "security" : [ {
    "access_token" : [ ]
  } ],
...
  "paths" : {
...
    "/api/data/domain/{domainId}" : {
      "post" : {
        "tags" : [ "data" ],
        "summary" : "Import a domain",
        "description" : "Import a domain by its id. The domain must be available in the configured import directory.",
        "operationId" : "importDomain",
        "parameters" : [ {
          "name" : "domainId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/UUID"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "string"
                }
              }
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "data" ],
        "summary" : "Clear a domain",
        "description" : "Clear a domain by its id. All nodes and relations will be removed.",
        "operationId" : "clearDomain",
        "parameters" : [ {
          "name" : "domainId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/UUID"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/data/domain/{domainId}/cancel" : {
      "post" : {
        "tags" : [ "data" ],
        "summary" : "Cancel a running import",
        "description" : "Cancel a running import by its domain id.",
        "operationId" : "cancelImport",
        "parameters" : [ {
          "name" : "domainId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/UUID"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "string"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/graph/domain" : {
      "post" : {
        "tags" : [ "graph" ],
        "summary" : "Create a domain",
        "description" : "Create a domain",
        "operationId" : "createDomain",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/DomainCreate"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Domain"
                }
              }
            }
          }
        }
      }
    },
    "/api/graph/domain/{domainId}/keys" : {
      "get" : {
        "tags" : [ "graph" ],
        "summary" : "Fetch a domain keys",
        "description" : "Fetch a domain keys",
        "operationId" : "fetchDomainKeys",
        "parameters" : [ {
          "name" : "domainId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/UUID"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "type" : "string"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/graph/domain/{id}" : {
      "get" : {
        "tags" : [ "graph" ],
        "summary" : "Fetch a domain",
        "description" : "Fetch a domain",
        "operationId" : "fetchDomain",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/UUID"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Domain"
                }
              }
            }
          }
        }
      },
      "put" : {
        "tags" : [ "graph" ],
        "summary" : "Update a domain",
        "description" : "Update a domain",
        "operationId" : "updateDomain",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/UUID"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/DomainUpdate"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Domain"
                }
              }
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "graph" ],
        "summary" : "Delete a domain",
        "description" : "Delete a domain",
        "operationId" : "deleteDomain",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/UUID"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/GraphDelta"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/graph/node" : {
      "post" : {
        "tags" : [ "graph" ],
        "summary" : "Create a node",
        "description" : "Create a node",
        "operationId" : "createNode",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/NodeCreate"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/GraphDelta"
                }
              }
            }
          }
        }
      }
    },
    "/api/graph/node/{id}" : {
      "put" : {
        "tags" : [ "graph" ],
        "summary" : "Update a node",
        "description" : "Update a node",
        "operationId" : "updateNode",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/UUID"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/NodeUpdate"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/GraphDelta"
                }
              }
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "graph" ],
        "summary" : "Delete a node",
        "description" : "Delete a node",
        "operationId" : "deleteNode",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/UUID"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/GraphDelta"
                }
              }
            }
          }
        }
      }
    },
    "/api/graph/relation" : {
      "post" : {
        "tags" : [ "graph" ],
        "summary" : "Create a relation",
        "description" : "Create a relation",
        "operationId" : "createRelation",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/RelationCreate"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/GraphDelta"
                }
              }
            }
          }
        }
      }
    },
    "/api/graph/relation/{id}" : {
      "put" : {
        "tags" : [ "graph" ],
        "summary" : "Update a relation",
        "description" : "Update a relation",
        "operationId" : "updateRelation",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/UUID"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/RelationUpdate"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/GraphDelta"
                }
              }
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "graph" ],
        "summary" : "Delete a relation",
        "description" : "Delete a relation",
        "operationId" : "deleteRelation",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/UUID"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/GraphDelta"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "Domain" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "$ref" : "#/components/schemas/UUID"
          },
          "name" : {
            "type" : "string"
          },
          "rootNodeId" : {
            "$ref" : "#/components/schemas/UUID"
          },
          "file" : {
            "type" : "string"
          },
          "columnMapping" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "string"
            }
          },
          "modelValid" : {
            "type" : "boolean"
          },
          "mappingValid" : {
            "type" : "boolean"
          }
        }
      },
      "DomainCreate" : {
        "type" : "object",
        "properties" : {
          "name" : {
            "type" : "string"
          },
          "rootNodeId" : {
            "$ref" : "#/components/schemas/UUID"
          },
          "file" : {
            "type" : "string"
          },
          "columnMapping" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "string"
            }
          }
        }
      },
      "DomainUpdate" : {
        "type" : "object",
        "properties" : {
          "name" : {
            "type" : "string"
          },
          "rootNodeId" : {
            "$ref" : "#/components/schemas/UUID"
          },
          "file" : {
            "type" : "string"
          },
          "columnMapping" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "string"
            }
          }
        }
      },
...
      "GraphDelta" : {
        "type" : "object",
        "properties" : {
          "trigger" : {
            "$ref" : "#/components/schemas/UUID"
          },
          "changedDomains" : {
            "uniqueItems" : true,
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/Domain"
            }
          },
          "removedDomains" : {
            "uniqueItems" : true,
            "type" : "array",
            "items" : {
              "format" : "uuid",
              "pattern" : "[a-fA-F0-9]{8}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{12}",
              "type" : "string"
            }
          },
          "changedNodes" : {
            "uniqueItems" : true,
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/Node"
            }
          },
          "removedNodes" : {
            "uniqueItems" : true,
            "type" : "array",
            "items" : {
              "format" : "uuid",
              "pattern" : "[a-fA-F0-9]{8}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{12}",
              "type" : "string"
            }
          },
          "changedRelations" : {
            "uniqueItems" : true,
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/Relation"
            }
          },
          "removedRelations" : {
            "uniqueItems" : true,
            "type" : "array",
            "items" : {
              "format" : "uuid",
              "pattern" : "[a-fA-F0-9]{8}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{12}",
              "type" : "string"
            }
          }
        }
      },
...
      "Node" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "$ref" : "#/components/schemas/UUID"
          },
          "label" : {
            "type" : "string"
          },
          "domainIds" : {
            "uniqueItems" : true,
            "type" : "array",
            "items" : {
              "format" : "uuid",
              "pattern" : "[a-fA-F0-9]{8}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{12}",
              "type" : "string"
            }
          },
          "color" : {
            "type" : "string"
          },
          "properties" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/Property"
            }
          }
        }
      },
      "NodeCreate" : {
        "required" : [ "label" ],
        "type" : "object",
        "properties" : {
          "label" : {
            "pattern" : "\\S",
            "type" : "string"
          },
          "domainIds" : {
            "minItems" : 1,
            "uniqueItems" : true,
            "type" : "array",
            "items" : {
              "format" : "uuid",
              "pattern" : "[a-fA-F0-9]{8}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{12}",
              "type" : "string"
            }
          },
          "color" : {
            "type" : "string"
          },
          "properties" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/Property"
            }
          }
        }
      },
      "NodeUpdate" : {
        "required" : [ "label" ],
        "type" : "object",
        "properties" : {
          "label" : {
            "pattern" : "\\S",
            "type" : "string"
          },
          "domainIds" : {
            "minItems" : 1,
            "uniqueItems" : true,
            "type" : "array",
            "items" : {
              "format" : "uuid",
              "pattern" : "[a-fA-F0-9]{8}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{12}",
              "type" : "string"
            }
          },
          "color" : {
            "type" : "string"
          },
          "properties" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/Property"
            }
          }
        }
      },
...
      "Property" : {
        "type" : "object",
        "properties" : {
          "name" : {
            "type" : "string"
          },
          "type" : {
            "$ref" : "#/components/schemas/PropertyType"
          },
          "primary" : {
            "type" : "boolean"
          }
        }
      },
      "PropertyType" : {
        "enum" : [ "int", "long", "double", "float", "string", "boolean", "date" ],
        "type" : "string"
      },
...
      "Relation" : {
        "required" : [ "type" ],
        "type" : "object",
        "properties" : {
          "id" : {
            "$ref" : "#/components/schemas/UUID"
          },
          "sourceId" : {
            "$ref" : "#/components/schemas/UUID"
          },
          "targetId" : {
            "$ref" : "#/components/schemas/UUID"
          },
          "domainIds" : {
            "uniqueItems" : true,
            "type" : "array",
            "items" : {
              "format" : "uuid",
              "pattern" : "[a-fA-F0-9]{8}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{12}",
              "type" : "string"
            }
          },
          "type" : {
            "pattern" : "\\S",
            "type" : "string"
          },
          "primary" : {
            "type" : "boolean"
          },
          "multiple" : {
            "type" : "boolean"
          },
          "color" : {
            "type" : "string"
          },
          "properties" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/Property"
            }
          }
        }
      },
      "RelationCreate" : {
        "required" : [ "type", "sourceId", "targetId", "domainIds" ],
        "type" : "object",
        "properties" : {
          "type" : {
            "pattern" : "\\S",
            "type" : "string"
          },
          "sourceId" : {
            "$ref" : "#/components/schemas/UUID"
          },
          "targetId" : {
            "$ref" : "#/components/schemas/UUID"
          },
          "domainIds" : {
            "minItems" : 1,
            "uniqueItems" : true,
            "type" : "array",
            "items" : {
              "format" : "uuid",
              "pattern" : "[a-fA-F0-9]{8}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{12}",
              "type" : "string"
            }
          },
          "primary" : {
            "type" : "boolean"
          },
          "multiple" : {
            "type" : "boolean"
          },
          "properties" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/Property"
            }
          }
        }
      },
      "RelationUpdate" : {
        "required" : [ "type", "domainIds" ],
        "type" : "object",
        "properties" : {
          "type" : {
            "pattern" : "\\S",
            "type" : "string"
          },
          "domainIds" : {
            "minItems" : 1,
            "uniqueItems" : true,
            "type" : "array",
            "items" : {
              "format" : "uuid",
              "pattern" : "[a-fA-F0-9]{8}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{12}",
              "type" : "string"
            }
          },
          "primary" : {
            "type" : "boolean"
          },
          "multiple" : {
            "type" : "boolean"
          },
          "properties" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/Property"
            }
          }
        }
      },
...
      "UUID" : {
        "format" : "uuid",
        "pattern" : "[a-fA-F0-9]{8}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{12}",
        "type" : "string"
      }
    },
    "securitySchemes" : {
      "access_token" : {
        "type" : "http",
        "scheme" : "Bearer"
      },
      "SecurityScheme" : {
        "type" : "openIdConnect",
        "description" : "Authentication",
        "openIdConnectUrl" : "http://localhost:8080/auth/realms/master/.well-known/openid-configuration"
      }
    }
  }
}